IncompressibleAdjoint.Equations

IncompressibleAdjoint.Equations.G_params — Method

G_params(Ω::Triangulation, params)

Compute the tensor G and the values GG and gg according to the VMS formulation proposed by [1]

IncompressibleAdjoint.Equations.adjoint_conservation — Method

adjoint_conservation(params)

Steady conservation equations of the adjoint problem

IncompressibleAdjoint.Equations.adjoint_steady_SUPG — Method

adjoint_steady_SUPG(params::Dict{Symbol,Any})

It provides the set of Adjoint Equations Symbol convention: O. Soto, & R. Lohner. (2004). On the Boundary Computation of Flow Sensitivities. https://doi.org/10.2514/6.2004-112 SUPG Stabilization: Srinath, D. N., & Mittal, S. (2010). An adjoint method for shape optimization in unsteady viscous flows. Journal of Computational Physics, 229(6), 1994–2008. https://doi.org/10.1016/j.jcp.2009.11.019

IncompressibleAdjoint.Equations.adjoint_steady_VMS — Method

adjoint_steady_VMS(params::Dict{Symbol,Any})

It provides the set of Adjoint Equations Symbol convention: O. Soto, & R. Lohner. (2004). On the Boundary Computation of Flow Sensitivities. https://doi.org/10.2514/6.2004-112

IncompressibleAdjoint.Equations.adjoint_unsteady_SUPG — Method

adjoint_unsteady_SUPG(params::Dict{Symbol,Any})

It provides the set of Adjoint Equations Symbol convention: O. Soto, & R. Lohner. (2004). On the Boundary Computation of Flow Sensitivities. https://doi.org/10.2514/6.2004-112 SUPG Stabilization: Srinath, D. N., & Mittal, S. (2010). An adjoint method for shape optimization in unsteady viscous flows. Journal of Computational Physics, 229(6), 1994–2008. https://doi.org/10.1016/j.jcp.2009.11.019

IncompressibleAdjoint.Equations.adjoint_unsteady_VMS — Method

adjoint_unsteady_VMS(params::Dict{Symbol,Any})

It provides the set of Adjoint Equations Symbol convention: O. Soto, & R. Lohner. (2004). On the Boundary Computation of Flow Sensitivities. https://doi.org/10.2514/6.2004-112 SUPG Stabilization: Srinath, D. N., & Mittal, S. (2010). An adjoint method for shape optimization in unsteady viscous flows. Journal of Computational Physics, 229(6), 1994–2008. https://doi.org/10.1016/j.jcp.2009.11.019

IncompressibleAdjoint.Equations.compute_G — Method

compute_G(trian::Gridap.Geometry.BodyFittedTriangulation, params)

Compute G (AbstractArray of TensorValues)

IncompressibleAdjoint.Equations.compute_GG — Method

compute_GG(trian::Gridap.Geometry.BodyFittedTriangulation, params)

Compute GG

IncompressibleAdjoint.Equations.compute_d — Method

compute_d(trian::Gridap.Geometry.BodyFittedTriangulation, params) #trian == Ω

The inverse of the cell-map-field. It is evaluted in the middle of the refernce domain.

IncompressibleAdjoint.Equations.compute_gg — Method

compute_GG(trian::Gridap.Geometry.BodyFittedTriangulation, params)

Compute gg

IncompressibleAdjoint.Equations.direct_differentiation_steady_SUPG — Method

direct_differentiation_SUPG(params::Dict{Symbol,Any})

Formulation from Janssens -P Vandenschrick -K Stevens -G Alessi, B. (n.d.). THE CONTINUOUS ADJOINT APPROACH APPLIED TO THE STABILIZED FINITE-ELEMENT FORMULATION OF THE INCOMPRESSIBLE NAVIER-STOKES EQUATIONS. www.euroturbo.eu

IncompressibleAdjoint.Equations.direct_differentiation_unsteady_SUPG — Method

direct_differentiation_unsteady_SUPG(params::Dict{Symbol,Any})

IncompressibleAdjoint.Equations.direct_differentiation_unsteady_VMS — Method

direct_differentiation_unsteady_VMS(params::Dict{Symbol,Any})

IncompressibleAdjoint.Equations.h_param — Method

h_param(Ω::Triangulation, D::Int64)

For a given triangulation Ω it computes the cell size $Area^{1/2}, D=2$ $Volume^{1/3}, D=3$

IncompressibleAdjoint.Equations.primal_steady_SUPG — Method

primal_steady_SUPG(params::Dict{Symbol,Any})

Navier-Stokes SUPG stabilized set of equations

IncompressibleAdjoint.Equations.primal_steady_VMS — Method

primal_steady_VMS(params::Dict{Symbol,Any})

Navier-Stokes VMS stabilized set of equations

IncompressibleAdjoint.Equations.primal_unsteady_SUPG — Method

primal_steady_SUPG(params::Dict{Symbol,Any})

Navier-Stokes unsteady SUPG stabilized set of equations

IncompressibleAdjoint.Equations.primal_unsteady_VMS — Method

primal_steady_VMS(params::Dict{Symbol,Any})

Navier-Stokes unsteady VMS stabilized set of equations

IncompressibleAdjoint.Equations.τb — Method

τb(u, h, ν::Real, dt::Real)

Stabilization parameter for SUPG formulation. Janssens, B. (2014). Numerical modeling and experimental investigation of ﬁne particle coagulation and dispersion in dilute ﬂows.

IncompressibleAdjoint.Equations.τc — Method

τc(uu, params::Dict{Symbol,Any})

Stabilization parameter for VMS formulation. Bazilevs, Y., Calo, V. M., Cottrell, J. A., Hughes, T. J. R., Reali, A., & Scovazzi, G. (2007). Variational multiscale residual-based turbulence modeling for large eddy simulation of incompressible flows. Computer Methods in Applied Mechanics and Engineering, 197(1–4), 173–201. https://doi.org/10.1016/j.cma.2007.07.016

IncompressibleAdjoint.Equations.τm — Method

τm(uu, params::Dict{Symbol,Any})

Stabilization parameter for VMS formulation. Bazilevs, Y., Calo, V. M., Cottrell, J. A., Hughes, T. J. R., Reali, A., & Scovazzi, G. (2007). Variational multiscale residual-based turbulence modeling for large eddy simulation of incompressible flows. Computer Methods in Applied Mechanics and Engineering, 197(1–4), 173–201. https://doi.org/10.1016/j.cma.2007.07.016

IncompressibleAdjoint.Equations.τsu — Method

τsu(u, h, ν::Real, dt::Real)

Stabilization parameter for SUPG formulation. Janssens, B. (2014). Numerical modeling and experimental investigation of ﬁne particle coagulation and dispersion in dilute ﬂows.