IncompressibleAdjoint.Geometry

IncompressibleAdjoint.Geometry.CST — Type

CST

It stores the weight for the upper and lower side of the airfoil

IncompressibleAdjoint.Geometry.ControlPoints — Type

It stores the coordinate of the spline control points.

IncompressibleAdjoint.Geometry.ShiftUpdate — Type

It stores the update for each degree of freedom of the the airfoil shape

IncompressibleAdjoint.Geometry.SplinePoints — Type

It provides the arifoil points to Gmsh to create the mesh. In tag are stored the "top" or "bottom" string for each point to identify if it is on the top or bottom surface

IncompressibleAdjoint.Geometry.CST_NACA0012 — Method

CST_NACA0012()

It gives the CST of the NACA0012 airfoil. t is the leading thickness.t=0means sharp.t = 0.0126` is a common value.

IncompressibleAdjoint.Geometry.ClassFunction — Method

ClassFunction(x::Vector{Float64},N1::Real,N2::Real)

Compute the class function $C = \phi^N1 \cdot(1-\phi)^N2$

IncompressibleAdjoint.Geometry.NACA00 — Method

NACA00(x; t=0.12)

Analytical function for a NACA00XX airfoil. t=0.12 computes the NACA0012 and so on.

IncompressibleAdjoint.Geometry.ShapeFunction — Method

ShapeFunction(w::Vector,x::Array{Float64})

Compute the shape function

IncompressibleAdjoint.Geometry.circle — Method

circle(x; diameter=1.0)

It return the equation of a circle with the center in [-diameter/2, 0.0].

IncompressibleAdjoint.Geometry.compute_cst_error — Method

compute_cst_error(w,p)

Compute the error of approximation

IncompressibleAdjoint.Geometry.compute_error — Method

compute_error(y0,y)

Compute the error of approximation. y0 are the original points, y are the new ones.

IncompressibleAdjoint.Geometry.create_msh — Method

create_msh(spline_points::SplinePoints; AoA=0.0, iter = 0, chord= 1.0, mesh_ref=1.0)

From a set of spline_points it creates the .msh file. Incresing mesh_ref is increasing the mesh density.

IncompressibleAdjoint.Geometry.get_idx_control_points — Method

Given a vector of the position of control points in control_points_position it provides the indexes of the points on the boundary

IncompressibleAdjoint.Geometry.get_radius_shift — Method

get_radius_shift(shift)

Given the shift, it computes the radius-influence of it.

IncompressibleAdjoint.Geometry.split_w_wlwu — Method

split_w_wlwu(w::Vector{Float64},split_idx::Tuple{Int64})

Creates 2 vector, one for wl and one for wu