IncompressibleAdjoint

Gridap.ODEs.ODETools.solve_step! — Method

Gridap.ODEs.ODETools.solve_step!(uf::AbstractVector,
solver::ThetaMethodBackw,
op::AffineODEOperator,
u0::AbstractVector,
t0::Real,
cache)

Support the new type ThetaMethodBackw which allows integration back in time, useful for unsteady adjoint problems

IncompressibleAdjoint.compute_airfoil_coefficients — Method

compute_airfoil_coefficients(uh,ph,nΓ,dΓ,params::Dict{Symbol,Any})

Compute the normaization of airfoil forces, obtaining CD and CL

IncompressibleAdjoint.compute_airfoil_forces — Method

compute_airfoil_forces(uh,ph,nΓ,dΓ,params::Dict{Symbol,Any})

It computes Drag and Lift over params[:tagname] boundary. It takes into account pressure and velocity gradient. It needs the normals pointings outward respect to the body.

IncompressibleAdjoint.compute_drag — Method

compute_drag(uh,ph,nΓ,dΓ,params)

Objective function minimizing the Drag

IncompressibleAdjoint.compute_lift — Method

compute_lift(uh,ph,nΓ,dΓ,params)

Objective function, it wants the CL to reach CLtarget. It return CL so it can be monitored.

IncompressibleAdjoint.compute_sensitivity — Method

compute_sensitivity(model::DiscreteModel, params::Dict{Symbol,Any}, uh0,ph0,ϕu0, ϕp0; objective_function=compute_drag)

From the solution of the primal flow uh0 ph0, and the adjoint flow ϕu0 ϕp0 it computes the senstivities according to the objective function. J2 contributions are splitted and the gradients computed individually to avoid numerical cancellation

IncompressibleAdjoint.obj_fun — Method

obj_fun(model::DiscreteModel, params::Dict{Symbol,Any}, uh,ph, fun)

Wrapper that evaluates fun the objective function. fun is a user-defined function. See eg. compute_drag, compute_lift functions. It return fitnessval,E. fitnessval is the function value that need to be minimized. E is a value that we want to monitor.

IncompressibleAdjoint.update_ũ — Method

update_ũ(ũ_vec::Vector)

It extimates the velocity field at the next time step using a Taylor-like expansion based on th previous 4 time steps.

IncompressibleAdjoint.verifykey — Method

verifykey(params::Dict{Symbol,Any},keyname; val = false)

It check if the dictionary params has the entry keyname. If not it adds the new entry with the value val. It is used to add default values